refactor : sync test folder with src file folder
