+++ tests/core/speech_two_ch1.wav
