+++ tests/core/speech_two_ch2.wav
