+++ tests/core/test_audio.py
