+++ tests/evaluators/test_single_stimulus_evaluator.py
